requirements/base.txt
Django==5.2.3
asgiref==3.8.1
Pillow==11.2.1

# Django 3rd party modules
django-allauth==65.9.0
django-filter==24.2
django-extensions==3.2.3
django-model-utils==5.0.0
djangorestframework==3.16.0
django-constance==4.3.2
django-easy-audit==1.3.7
django-widget-tweaks==1.5.0
django-celery-beat==2.8.1
django-sesame==3.2.2
django-formtools==2.5.1
django-cities-light==3.10.2
django-autocomplete-light==3.12.1

# Environment settings
python-dotenv==1.1.0
python-decouple==3.8
dj-database-url==3.0.0

# Celery
redis==5.2.1
celery==5.5.0

# Others
requests==2.32.3
setuptools==70.0.0
